Fixed duration parsing in LSFCommand (as before in PBSCommand) and refactored out common code.
